# gh-121109: Fix performance of tarfile reading with "r|*"